pyproject.toml
[project]
name = "instagram-dm-mcp"
version = "0.1.0"
description = "Instagram DM MCP server and CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.8",
    "instagrapi>=2.3.0",
    "mcp>=1.9",
    "pillow>=8.1",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project.scripts]
instagram = "src.cli:main"
